+++ src/app/advertiser/campaign-details/ad-list/ad-list-item/ad-list-item.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { Router } from '@angular/router';
 import { HttpErrorResponse } from '@angular/common/http';
-import { MatDialog } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
 import { Store } from '@ngrx/store';
 import { AdvertiserService } from 'advertiser/advertiser.service';
 import { adStatusesEnum } from 'models/enum/ad.enum';