+++ src/app/common/common.module.ts
@@ -5,24 +5,19 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms'
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations'
 import { MatSelectModule } from '@angular/material/select'
 import { MatChipsModule } from '@angular/material/chips'
-import {
-  MatButtonModule,
-  MatCheckboxModule,
-  MatDialogContent,
-  MatDialogModule,
-  MatIconModule,
-  MatInputModule,
-  MatMenuModule,
-  MatSlideToggle,
-  MatSnackBarModule,
-  MatSpinner,
-  MatToolbarModule,
-  MatTooltipModule,
-} from '@angular/material'
+import { MatButtonModule } from '@angular/material/button'
+import { MatCheckboxModule } from '@angular/material/checkbox'
+import { MatDialogContent, MatDialogModule } from '@angular/material/dialog'
+import { MatIconModule } from '@angular/material/icon'
+import { MatInputModule } from '@angular/material/input'
+import { MatMenuModule } from '@angular/material/menu'
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner'
+import { MatSlideToggle, MatSlideToggleModule } from '@angular/material/slide-toggle'
+import { MatSnackBarModule } from '@angular/material/snack-bar'
+import { MatToolbarModule } from '@angular/material/toolbar'
+import { MatTooltipModule } from '@angular/material/tooltip'
 import { MatDatepickerModule } from '@angular/material/datepicker'
 import { MatMomentDateModule } from '@angular/material-moment-adapter'
-import { MatSlideToggleModule } from '@angular/material/slide-toggle'
 import { MatExpansionModule } from '@angular/material/expansion'
 import { ChartsModule } from 'ng2-charts'
 import { ClickOutsideDirective } from './clickOutside.directive'
@@ -157,9 +152,6 @@ const appComponents = [
     ...dialogs,
     ...appComponents,
   ],
-  entryComponents: [
-    ...dialogs,
-  ],
   providers: [
     AssetHelpersService,
     BidStrategyService,
@@ -171,12 +163,11 @@ const appComponents = [
     ...appComponents,
     BrowserAnimationsModule,
     MatTooltipModule,
-    MatSpinner,
     MatDialogContent,
+    MatProgressSpinnerModule,
     MatSlideToggle,
     MatIconModule,
   ],
 })
-
 export class AppCommonModule {
 }