+++ src/app/common/dialog/change-automatic-withdraw-dialog/change-automatic-withdraw-dialog.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
-import { MatDialogRef } from '@angular/material';
+import { MatDialogRef } from '@angular/material/dialog';
 import { Store } from '@ngrx/store';
 
 import { SettingsService } from 'settings/settings.service';