+++ src/app/publisher/dashboard/dashboard.component.ts
@@ -4,14 +4,13 @@ import * as moment from 'moment';
 
 import { ChartService } from 'common/chart.service';
 import { ChartComponent } from 'common/components/chart/chart.component';
-import { SiteListComponent } from 'publisher/site-list/site-list.component';
 import { HandleSubscription } from 'common/handle-subscription';
 import { Site, SitesTotals } from 'models/site.model';
 import { ChartFilterSettings } from 'models/chart/chart-filter-settings.model';
-import { ChartData } from 'models/chart/chart-data.model';
-import { ChartLabels } from 'models/chart/chart-labels.model';
+import { ChartDataSets } from 'chart.js';
+import { Label } from 'ng2-charts';
 import { AppState } from 'models/app-state.model';
-import { createInitialArray } from 'common/utilities/helpers';
+import { createInitialDataSet } from 'common/utilities/helpers';
 
 import * as publisherActions from 'store/publisher/publisher.actions';
 import { appSettings } from 'app-settings';
@@ -27,7 +26,6 @@ import { reportType } from 'models/enum/user.enum';
 })
 export class DashboardComponent extends HandleSubscription implements OnInit {
   @ViewChild(ChartComponent) appChartRef: ChartComponent;
-  @ViewChild(SiteListComponent) siteListRef: SiteListComponent;
 
   sites: Site[];
   sitesLoaded: boolean = false;
@@ -37,8 +35,8 @@ export class DashboardComponent extends HandleSubscription implements OnInit {
   barChartValue: number;
   barChartDifference: number;
   barChartDifferenceInPercentage: number;
-  barChartLabels: ChartLabels[] = [];
-  barChartData: ChartData[] = createInitialArray([{data: []}], 1);
+  barChartLabels: Label[] = [];
+  barChartData: ChartDataSets[] = createInitialDataSet();
 
   currentChartFilterSettings: ChartFilterSettings;
 
@@ -88,7 +86,7 @@ export class DashboardComponent extends HandleSubscription implements OnInit {
       .pipe(take(1))
       .subscribe(data => {
         this.barChartData[0].data = data.values;
-        this.barChartData[0].currentSeries = chartFilterSettings.currentSeries.label;
+        this.barChartData[0].label = chartFilterSettings.currentSeries.label;
         this.barChartLabels = data.timestamps.map(item => moment(item).format());
         this.barChartValue = data.total;
         this.barChartDifference = data.difference;