+++ src/app/publisher/publisher.module.ts
@@ -1,20 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { ChartsModule } from 'ng2-charts';
-import {
-  MatAutocompleteModule,
-  MatButtonToggleModule,
-  MatCheckboxModule,
-  MatExpansionModule,
-  MatFormFieldModule,
-  MatInputModule,
-  MatOptionModule,
-  MatPaginatorModule,
-  MatRadioModule,
-  MatSelectModule,
-  MatTabsModule,
-} from '@angular/material';
+import { MatAutocompleteModule } from '@angular/material/autocomplete';
+import { MatButtonToggleModule } from '@angular/material/button-toggle';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatOptionModule } from '@angular/material/core';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatRadioModule } from '@angular/material/radio';
+import { MatSelectModule } from '@angular/material/select';
+import { MatTabsModule } from '@angular/material/tabs';
 import { PublisherRoutingModule } from './publisher-routing.module';
 import { PublisherGuard } from './publisher-guard.service';
 import { AppCommonModule } from 'common/common.module';
@@ -37,8 +34,7 @@ import { ClassifierListItemComponent } from './classifier/classifier-list-item/c
 import { SiteResolver } from './resolvers/site.resolver';
 import { FilteringCriteriaResolver } from './resolvers/filtering-criteria.resolver';
 import { AdUnitSizesResolver } from './resolvers/ad-unit-sizes.resolver';
-import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { library } from '@fortawesome/fontawesome-svg-core';
+import { FaIconLibrary, FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { fas } from '@fortawesome/free-solid-svg-icons';
 import { ClassifierFilteringComponent } from 'publisher/classifier/classifier-filtering/classifier-filtering.component';
 import { MatchingBannerSizesResolver } from 'publisher/resolvers/matching-banner-sizes.resolver';
@@ -54,8 +50,6 @@ import {
   MetaverseInstructions
 } from 'publisher/edit-site/edit-site-basic-info/metaverse-instructions/metaverse-instructions.component'
 import { SiteOptionsResolver } from 'publisher/resolvers/site-options.resolver'
-
-library.add(fas);
 
 
 const editSiteComponents = [
@@ -102,7 +96,6 @@ const publisherComponents = [
     MatSelectModule,
     MatExpansionModule,
     MatRadioModule,
-    ChartsModule,
     FontAwesomeModule,
     MatTabsModule,
   ],
@@ -119,12 +112,10 @@ const publisherComponents = [
     ...publisherComponents,
     ...editSiteComponents,
   ],
-  entryComponents: [
-    SiteCodeDialogComponent,
-    SiteCodeCryptovoxelsDialogComponent,
-    SiteCodeMetaverseDialogComponent,
-  ]
 })
 
 export class PublisherModule {
+  constructor (library: FaIconLibrary) {
+    library.addIconPacks(fas)
+  }
 }