+++ CHANGELOG.md
@@ -7,17 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.18.0] - 2022-06-02
+### Changed
+- Angular 13
+### Fixed
+- Setting date range for charts
+- User label when no e-mail
+
 ## [1.17.1] - 2022-05-30
 ### Fixed
 - Newsletter subscription
 - DCL builder link
 - Edit site's option for manual approval
 
+### Added
+- Link to wiki guide on how to add an HTML file 
+
 ## [1.17.0] - 2022-05-19
 ### Added
 - Links to publisher sites in publisher panel
 - Configuration default site classification and filtering options on admin panel
-- Opportunity for admin to ban unban and delete users 
+- Opportunity for admin to ban, unban and delete users
 ### Changed
 - Wallet login buttons visibility
 - Angular 7
@@ -417,7 +427,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Publisher features (Sites & AdUnits)
 
  
-[Unreleased]: https://github.com/adshares/adpanel/compare/v1.17.1...develop
+[Unreleased]: https://github.com/adshares/adpanel/compare/v1.18.0...develop
+[1.18.0]: https://github.com/adshares/adpanel/compare/v1.17.1...v1.18.0
 [1.17.1]: https://github.com/adshares/adpanel/compare/v1.17.0...v1.17.1
 [1.17.0]: https://github.com/adshares/adpanel/compare/v1.16.1...v1.17.0
 [1.16.1]: https://github.com/adshares/adpanel/compare/v1.16.0...v1.16.1